requirements.txt
fastapi
uvicorn
pydantic
python-multipart
aiofiles
faiss-cpu
sentence-transformers
llama-cpp-python
pdfplumber
sqlalchemy
asyncpg
python-dotenv
tqdm
requests
transformers>=4.41.0
accelerate
